requirements.txt
streamlit
mira-sdk
requests
python-dotenv
# ...other dependencies...
streamlit-extras
streamlit-card 
# ...existing dependencies...